stable/create-kernel-tasks: Stop unlinking old TBs
When creating a new tracking bug which replaces a previous one of the
same cycle, we remove the old tracking bug as a derivative from the
(old) master. Since we stopped updating forward pointers, this got
reduced to erase the master bug pointer from the old tracking bug.
However this is unnecessary as the old tracking bug will no longer be
considered as soon as it gets duplicated. Also this removes information
which might be needed again in case the duplication has to be undone.

Therfor completely drop any code which removes the old tracking bug as a
derivative.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>