stable/create-kernel-tasks: Stop duplication of tracking bugs
While this initially sounded like a good idea (and reflected what we
were manually doing) it became more of a problem lately. The reason is
that re-spins happen much more often these days. And even if working on
the new kernels immediately after creating the new tracking bugs, it
takes something like 8 hours (rather 24 hours if a re-spin includes a
master kernel) until the new binaries are in proposed.

And since duplicated tracking bugs no longer show up, this takes away a
lot of time for testing which could and should still be ongoing. The
re-spins should handle only a very specific portion of change which
(apart from the completely unbootable case) means the test results so
far are still meaningful.

For that reason, the duplication should wait until the very last moment
(when the new kernel is ready to go into proposed and only stops because
of the previous kernel there). Maybe even longer (until the new kernel
is in proposed) to be able to handle problems which happen during that
copy process.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>