admin: admin-lp-git-permission -- add support for adding cranker repo subscriptions
For private git repositories we want crankers to be able to read the
source but not push to it.  Rather than adding them to the primary team
which would make them admins of the repository we can add a repository
subscription for them, this gives them access to the code.  We request
no email updates on this subscription.

Signed-off-by: Andy Whitcroft <apw@canonical.com>