swm: bug -- replaces duplicate tracker when idle
Add support for marking trackers as replacing the tracker in a previous
spin.  This allows duplicates to be marked in the new trackers before
they are active, and obviates the need to touch the replaced tracker
which may fail due to overlapping updates by swm.  The proposed
replacement is indicated via a replaces key in the new tracker's
swm-properties.

The intent here is to allow trackers to be left open until the new
tracker is sufficiently progressed to replace it.  However, when
the target trackers is not adding value (for example allowing other
testing to continue) then we should dup it through immediatly.

Add support for detecting and applying replaces.  Use this when
the target is deemed inactive (not in testing currently).

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>