ktl/tracking_bug2.py: handle multiple trackers per source
When we are performing a respin it is entirely possible to have the old
spin in testing and the new building in the build PPA.  In this scenario
the trackers: list on our parent will contain multiple trackers for
that source:

    linux-oem: bug 123, bug 234

We really only care about the latest one in this scenario.  As launchpad
bug numbers are monotonic increasing use the highest bug as the valid one.

Signed-off-by: Andy Whitcroft <apw@canonical.com>