gen-sru-announce: use SruCycle to determine cycle length
When emitting data for a cycle use SruCycle to determine its length.
Form a layout from the cycle start/end dates.  Patches due the wednesday
before.  Prep week is the first 5 working days.  Testing is the
remainder of the cycle stopping on the last friday.  Release is the
final Monday.  Use this layout to form the cycle in text and html.

Signed-off-by: Andy Whitcroft <apw@canonical.com>