swm: package,check_component -- use all archives in a route
When accumulating route from kernel-series record all of the offered
routes rather than just the first one.  When looking for confirmation of
publication use all possible routes.  This allows us to use ppa and ppa2
for building kernels.

When scanning a route use all of the listed archives.  If we find the
exact version we are looking for in an archive/pocket use that one.
When there is no match use whatever we found in the first listed archive.
Finally correct a few consumers who really only care about the first
route to handle the switch to a list.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>