kernel-sru-dashboard: report debs in:ppa when building
We are currently only reporting a kernel as in:ppa when it is fully
built and ready to copy.  This makes it not in:ppa if it is ready but its
parent is not ready.  Actually we care if it is in the ppa from a version
is now committed point of view, so actually whenever it is building is
sufficient to display in:ppa.  This is indicated by prepare-package being
Fix Committed or later; switch to that.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>