cranky: review -- simplify the diffstat to make it more comparible
It is useful to be able to compare the diffstat on different reviews.
To this end fiddle with the git diff options such that the stat component
of the diff does not truncate the filenames based on width which makes
them unpredicable for different length base paths.  Limit the graph
width else it will fill the available space (and become huge).

Signed-off-by: Andy Whitcroft <apw@canonical.com>