ktl/kmsgq: add a service specific message queue initialiser
We mostly wish to instantiate queue end points into a specific service.
Add a new interface which takes the service name as a parameter.
This allows us to use that to find and fill defaults for things like AQPM
server and credentials for it.  For now these are all hard wired.

Signed-off-by: Andy Whitcroft <apw@canonical.com>