stable/create-sru-cards: Add configurable list targets
Mainly adds the possibility to have a non-default list destination when
adding new cards. This way we can populate any list and not just a
default one.

While at it move some dry-run handling into the board class. That way it
is possible to get a bit more coverage when testing in dry-run mode.

Also rework the yaml template to have same new tasks, more information
in some, and spread things to various lists. I am no longer filling the
"Backlog" list so the crank cards which are automatically moved are no
longer hiding the tasks which have to be manually moved.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>