swm: manager -- only trigger dependants on master change
Currently in --dependants mode we rescan every dependant (deriviatives
etc) of a tracker every time we scan that tracker.  Often we will scan a
tracker and find nothing of interest has happened; for instance if we are
triggered for an amd64 publication against that tracker but this is not
the last architecture to complete then the scan was a noop.  Switch to
tracking the scanned and modification times for all trackers.  When we
complete a tracker scan run use this data to identify any additional
trackers which might now be affected and only scan those.  We store this
data in the status.yaml to allow it to be maintained between runs.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>