swm: snap -- only request fields which are valid
The store is about to start validating our requests.  It seems we are
requesting a heap of fields (because we want them) which are always
included and are therefore invalid to request.  Store contacts recommended
only requesting revision,version to get the pieces we require, confirmed
that requesting this is sufficient to obtain the data we require.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>