mainline-builds: use correct all architecture and only build valid ones
We now are dropping i386 in eoan so we cannot actually build this arch for
newer mainline builds.  Firstly switch to using amd64 for arch all (which
we should have done way back in xenial) and then use the build control
files to limit builds to architectures the packaging actually supports.

Signed-off-by: Andy Whitcroft <apw@canonical.com>