swm: package,wft/prepare-packge -- primary package dwell till built
We would rather (as a policy) not start preparing derivatives until the
top-level primary kernel main packages are built as a failure there can
lead to a full respin of any derivatives.  Hold the prepare-package for
each of those back in Fix Committed until the build is complete.

This can be overridden by moving the prepare-package task to Fix
Released manually.  We will not attempt to move it back.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>