kernel-stable-dashboard: security-signoff should now be displayed
We have long only shown the security-signoff tasks for the primary
kernels.  Security have asked we no longer copy this status to
non-backports kernels such that all other derivatives now need to be
worked; they were doing so without the benefit of tasks to track it.
Show all security-signoff tasks.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>