stable/create-kernel-tasks: Make --spin required
Initially it sounded like a good idea to automatically pick the next
free spin number when it is not provided. However, it rather created
confusion as it is kind of required when adding tasks in a loop over
the supported series.

Now trying to see whether making this non-optional works better.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>