swm: manager -- add tracker fix framework
Some changes to a tracker can only safely be applied while we have the
lock for that bug.  The easiest way to do that is trigger scanning of that
bug and apply the fixes when we encounter the bug.  Add a framework to
detect and apply such fixes expressed as fix-* markers in the persistent
state for the tracker.  Do this atomically under the status lock so that
we will not remove the request to fix before the change is committed to
the tracker itself.

Note that no fixes are implemented yet.  Fix implementations are in
subsequent patches.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>