kernel-stable-dashboard: fix URL insertion once and for all
We have been mangling the produced HTML to insert links over the various
testing status elements for some time.  This is incredibly fragile and
has had to be fixed a number of times.  Rather than trying to fix it
yet again after the fact add proper support for applying the URLs as we
generate the blocks.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>