info/kernel-series: 20.04 should be a LTS
SRU Testing queries the kernel-series.yaml to derive if the
series under test is a LTS or not. If LTS is true, then
most clouds will append a '-lts' to the image name. Focal
does not have this tag which is breaking a few things in
the cloud provisioning process.

Signed-off-by: sfeole <sean.feole@canonical.com>