swm: swm-trello-sync -- hold off moving lost+found candidates for 30m
When we create new trackers we create the cards first.  This leads the
shover to move them to lost+found and the back to backlog when the
tracking bug itself comes into existance.  Avoid this pointless activity
by holding off moving lost+found candidates until they have been in
existance for 30m.  Note that as we have no creation date information we
approximate this by holding off touching the card if it was modified in
the period.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>