cranky/cranky-update-dependent: Add option to ignore the ABI check
Add a new command line --ignore-abi-check to force
cranky-update-dependent to continue even if the new ABI number is not
higher then the current one.

This is useful when re-spinning a kernel where only the upload number
should be increased. Also, in this scenario, the in-tree
update-version script usually produces a new changelog version without
any entries. Workaround that by including a generic changelog entry
and showing a warning to the user to review the changelog.

Acked-by: Andrea Righi <andrea.righi@canonical.com>
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>