ktl/kernel-series: Add variants field to sources
We are introducing the idea of variants produced by kernel source
packages. Examples of variants are -hwe-18.04(-edge), -- (for the
primary variant), and -wip (work-in-progress). Tracking variants
in kernel-series will allow cranky-fix to update the variants in
source packages in concert, and importantly updating the variant
in -meta without doing so in the kernel package (useful for
transitioning hwe-edge to hwe by updating only meta).

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>