cranky: crl/handle -- Always take package name from topmost changelog entry
When changing the source package name it's reasonable to do so
by opening a new entry with 'cranky open' and changing the name
for that entry. However crl will skip over the unreleased entry
and get the name from the next entry. This is reasonable for the
series, since it would be 'UNRELEASED', but unnecessary for the
package name.

Always take the source package name from the topmost entry to
support this sort of workflow.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>