cranky: update-snap -- follow copy-forward backwards
When we are in the early phase of a new development we copy kernels forward
but we wish to continue to provide snaps into the up-coming ucNN release.
At this point we are building snaps in the new development release chroots
but wanting to create snaps against the now-stable release.  Add support
for following the copy-forward directives back to find the snaps.  Also use
this information to determine where to look for the appropriate version
number; where the copy forward has not yet occured we will fail the build
rather than building the wrong version.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>