cranky: update-snap -- allow KERNEL_SOURCE to be a source handle
We do not have the base series in the URL for uc20 so allow it to be
specified locally in the snap definition so that we can find the object
in kernel-series.  Allow KERNEL_SOURCE to be a cranky style handle
(bionic:linux).  Retain compatibility with the KERNEL_SOURCE as a simple
source package name (linux-raspi2).  Continue to fallback to intuiting
the series name from the source git URL and the source from that and the
snap name when necessary.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>