swm: manager -- catch direct state modifications on trackers
When a task is directly set to a new state we do not internally see
that as a modification.  Look at the bug modification time and if it is
newer than our scan time assume it is modified.  This is tickled by our
own swm-properties updates but it is conceivable we could be waiting on
those too.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>