swm: swm-trello-sync: move cards to the bottom of the list when appropriate
We are interested in keeping tasks in the order they came ready in
the pending lists, but for the In Progress and Done lists we are more
interested in the latest changes.  Drop things onto the top of the In
Progress and Done lists and the bottom of the remainer.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>