swm-publishing: MonitorLaunchpadArchive handle publishing race
We detect publication changes by asking for all of the new records
since a specific date representing our last scan, they are returned in
reverse cronological order (newest to oldest).  This is an essential
part of the API because the list can be extended while we are iterating.

Generally entries added after our scan starts will appear chronologically
later than the first entry we have seen in this run.  This triggers
potential duplicates in the results from an ongoing scan as we move
from page to page.  This occurs because we start the second page at the
75th entry of the now extended list, ie at the 74th entry in our list.
This leads us to see that 74th entry twice once in each page; however no
entries are missed.  Entries can also be allocated and assigned but not
committed for some considerable time (think minutes); at the point that
the entire operation completes launchpad does a transactional commit and
the now old entry will appear; we could have scanned the period in which
this is places completely and thus can we might miss entries.

We will (already) handle the first scenario as we simply add each detected
trigger to the result set which nativly deduplicates the result.  For the
second, when we perform a scan across the time range in which the entry
will appear we will not see this entry and then never revisit the range,
leading to loss of the trigger entirely.  To handle this case always scan
at least 1 hour back from the current clock time.  This necessarily will
lead us to see duplicates between runs.  Add a replay array to each archive
which lets us detect and ignore entries we have successfully triggered.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>