swm-publishing: spread refresh based scanning across our scan window
We want to scan all of our live bugs at least every 4 hour scan window.
We want to do this to ensure that if we miss a triggering event we will
catch it within a reasonable time scale.  We also want to spread these
refreshes out as evenly as possible to ensure we do the smallest possible
scans in any single run (as the longer the run the higher the possibility
of failure of the batch).  In the face of a long scanning outage we also
want to try and catch up and not just scan them all in the following
4 hours.

To do this we look at the pending backlog, those trackers which were
last scanned more than 3.5 hours ago, and schedule enough into this batch
such that we would drain the current backlog in 1 hour.  This effectively
averages the loads in the hour window.  Applying this iterativly in each
scan run will refresh each tracker on average every 4 hours.  It will
also tend to spread the trackers out into the half hour each side of
the current 4 hour mark.  Over time this leads the trackers to even out
throughout this period.  When a required scan refreshes a large number
trackers they will form a spike, which will again average out over a
full hour the next cycle, and further in each later cycle.

Signed-off-by: Andy Whitcroft <apw@canonical.com>