cranky: shell-helper -- fix tree-main-path for working with unstable
When working with unstable it's common to have source trees
checked out in paths with "unstable" substituted for the series
name. Cranky supports this fine for most cases, however the
tree-main-path subcommand for shell-helper will return the path
to the development series kernel tree instead of the unstable
tree.

Change tree-main-path to look up the handle set and iterate the
trees in that object, similarly to how it is done for
source-packages-path. This gives the correct result for the
unstable use case.

Acked-by: Andrea Righi <andrea.righi@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>