ktl/tracking_bug2: switch to a task exclusion model for stable/devel
We currently have a random inclusion list for development.  This only
supports variant:debs tasks so you cannot actually generate a
variant:snap-debs tracker in development.

Actually there is little difference in the nominally valid tasks between
stable and development; we only want promote-to-release in development
and only promote-to-{updates,security} in stables, and some testing
and signoff tasks are elided in development.  Encapsulate these
directly.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>