swm: snap -- completely missing tracks are not populated
When checking if a snap is published into a new track where there is
currently nothing published we will incorrectly believe that the revision
(None) is equal to the non-existant build revision None and indicate the
snap is present and published.  Later when they start to come through we
will pull it back.  Detect this corner case and treat it as not present.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>