cranky/master-helpers/open: Handle new releases
If the ABI directory is not found and the current version ends in .0 it
means it's a dummy entry of a new release which doesn't have an ABI. In
that case, use the version of the previous changelog entry.

While at it, emit and error message and bail out if no ABI directory is
found.

Signed-off-by: Juerg Haefliger <juergh@canonical.com>