cranky: fix -- enhance dkms-build update support
We have added a new file-downloader helper that knows how to handle
librarian downloads such that they will work reliably on the ports
architectures.  Add support for copying this as part of syncing the
dkms-build family.  Also add support for switching the Build-Depends
for dkms-builds from wget to curl.  Enable all of this only when
file-downloader is present in the master kernel so we can apply same
progressivly.

Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>