swm: swm-trello-sync -- items in Backlog should not be in In Progress
We allow items to move ahead of their expected state sometimes, for example
you might pick a card out of On Deck and move it to In Progress and slap
your face on it before swm has any idea you are touching it (as that is
the interlock for owning the crank).  Currently we allow items destined
for Backlog to be in In Progress as well, this is not so reasonable,
if it isn't ready to execute executing it is bad period.  Drop this.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>