deploy: add a new deployment script to pull down needed dependencies
We have an external dependency which is needed for a few of the commands
which interact with trello.  Added a new deploy script which is intended
to prepare the repository for active use of its contents.  This will
pull down the external dependencies and link them into place.
Additionally this script can be used in --update mode from cron to keep
the primary tree and its depedencies all up to date.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>