swm: bug -- add support for non-task reasons
We have some phases of the life-cycle of a tracker which has no associated
actions.  For example prepare-package completes when the package is
uploaded but the next action promote-to-proposed does not officially
start until they are built and published.  It would be nice to report
things like this against its own synthesised task.  Add support for
reasons not associated with launchpad tasks.  At the same time convert
the special case 'overall' status to a synthetic task.

Signed-off-by: Andy Whitcroft <apw@canonical.com>