ktl/tracking_bug2.py: fix snap task creation
When deciding whether a task is valid for a snap tracking bug, the
decision is wrongly made based only on the primary snap for that given
series/source. Fix it by taking into consideration the snap name
provided by the caller.

While at it, update the pydoc of the parameters for
TrackingBugs.create().

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>