swm: bug -- if a duplicate throws a WorkflowBugError ignore it
In order to determine if the proposed pocket is clear we have to
instantiate any of our duplicate bugs and see if they own the contents.
Whe we instantiate them and they throw a WorkflowBugError (say they have
bad package information due to a kernel-series change) then we will die
and missattribute the error to the primary bug.  Catch these, report them
for the logs but then ignore them en-toto.  This is safe as we may
blocked when we should not but not release when we should not.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>