swm-publishing: add a 1hr timeout
When we are talking to launchpad to look things up sometimes the world
gets sad and we simply hang waiting.  Rather than try and handle this
in the individual transactions take advantage of the fact on error we
continue the next run as if nothing occured and just abort.  Add a
timeout of one hour to the swm-publishing-run script.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>