verify-release-ready: fix tag verification
We have evolved how we tag the closing commits by following a more
standand pattern which is done by 'cranky tag'. The verification done by
verify-release-ready was not updated and is still checking for some old
patterns. Update it to construct the expected tag name using the same
way that 'cranky tag' does it.

Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>