cranky: chroot: Install riscv64 cross toolchain and libc6-dev
Focal is gaining riscv64 architecture support, therefore we need
to install cross toolchain and libc6-dev packages in our chroots
to support updating configs. These packages are available for
bionic and later, and there's not any nice way to test for "focal
and later," so these packages are installed for all releases
except trusty and xenial.

Acked-by: Juerg Haefliger <juergh@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>