cranky: test-build: do not fail if test branch exists
The script fails if the test branch already exists. This prevents
paralell execution of the script. Fix it by checking if the branch
already exists and just continues in that case. The test branch is always
created using the commit sha1 so it's guaranteed that we are testing the
correct commit. The test commit will be deleted by the first script
process that finishes the execution.

Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>