adt-matrix: fix local seed pollution when adt-seed-accumulate does not supply
When we have no adt-seed-accumulate seed for a package (such as before
britney sees it) then we make up a seed using the linux package in the
series plus our package.  This uses the same copy of those seeds in each
iteration leading to significant pollution of later entries.  Make a
fresh copy for each iteration.

Signed-off-by: Andy Whitcroft <apw@canonical.com>