kernel-series.yaml: Initial fix for testing bionic/hwe
The testing section should list the flavours which get produced
or rather the name which maps to a kernel flavour in meta).
For hwe we have genric and lowlatency (virtual, too but that is
identical to generic).
While doing that also declare the correct variant (-hwe-18.04)
which currently is coded into the meta package rules file.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>