swm: package -- remove __get_published_sources
__get_published_sources has four rules for looking up sources handling all
combinations of abi supplied/not-supplied and pocket supplied/not-supplied.
However, The caller always supplies the pocket so rules 2 and 4 are never
used and are now dead code.  Separatly we have stopped supplying release
(package version) in order to record non-matching versions in pocket for
pocket_clear safety checks.  This makes rules 1 and 3 equivalent.  Pull
up the single valid rule into __is_fully_built.

Signed-off-by: Andy Whitcroft <apw@canonical.com>