swm: package -- switch to fixed dependent package version tracking
Dependent packages can have different versioning from the main package.
Meta and lbm both have versionin matching only the kernel's ABI, other
packages can have re-upload subscripts (+N +signedN etc).  When these
differ from the primary we currently use sloppy matches to find them and
always assume any possible package is the right one.  This makes it
impossible to track updates in these packages as a respin.

Move to tracking the specific versions of dependent packages we are
expecting.  Store these expected versions in the swm-properties for the
tracker so that they remain stable for the life of the tracker (we may
choose to update these manually to facilitate a dirty respin).  Where these
are not known determine these versions from the tags used on the git
repository, use similar sloppy matching to identify the initial version.

Note we are intentionally switching from matching on upload versions
to tag versions.  This is done for two reasons.  Firstly semantically
the tags are earlier in the life-cycle of the upload and later under
automation 2.0 we will only supply tags.  Secondly source-only uploads
(such as disco:linux) do not have uploads.

Signed-off-by: Andy Whitcroft <apw@canonical.com>