swm: wfl/prepare_package -- check for package tags and uploads together
Now that we have explicit version tracking we can use that to tell us
when we have any tags or uploads present.  When we do start reporting
missing tags and missing uploads.  Do this as a combined reason as
people may upload new packages and push the tags in either order.

Signed-off-by: Andy Whitcroft <apw@canonical.com>