devel: Add initial tools for created devel boards and cards
Add an initial version of scripts to create boards for tracking
of tasks associated with a specific upstream kernel. These are
based off of the stable tools, and like them they use yaml files
to describe the boards and cards to be created.

There's probably some of room here for creating libraries to
handle things in common with the stable scripts, but let's hold
off on that until we've worked out exactly what we want the devel
scripts to be doing.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>