stable/create-sru-cards.yaml: update description of pre-cycle test
The 'Get pre-cycle boot tests done' mentions that a dirty build should
be done for a quick boot test over the weekend before the cycle
officially starts. We never implemented the dirty build tests and we
are instead uploading real cranks of the main kernels for the weekend
test. Update the card description accordingly.

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>