swm: wft/workflow -- report building in all pockets consistently -- fix all_missing detection
We attempt to elide the build status for all packages when no packages
are present.  This prevents us listing literally everything as missing
as soon as we have a version.  However we do not correctly handle the
case where a package is not in failure.  If we upload just linux it will
short linux:building, linux-meta:missing etc.  But once linux completes,
we revert to thinking all packages are missing and elide the reporting
of the awol packages.

Switch to using the overall count of sources against missing failures.

Signed-off-by: Andy Whitcroft <apw@canonical.com>