mainline-builds: worker -- use routing keys to direct commands
Currently we simply bind to the name of the queue.  This effectivly
subscribes us to all messages on the queue, thus we have to support all
messages destined for the queue.  In preparation for having different
workers support different command lists switch to using routing keys.
Bind the queue to the routing key for each command we support
(todo.<cmd>).  Also subscribe our direct queue to the same routing keys
prefixed with a direct.<id>.  This allows us to direct any command to a
specific instance or send them to the queue for fair processing.

Note that this removes the separation between direct commands (called
admin commands in the UI) and pool commands.  We have long wanted to be
able to direct any command to a specific unit for example when testing
new code.  No safety was added by limiting some commands to the direct
interface.

Signed-off-by: Andy Whitcroft <apw@canonical.com>