cve-autotriage: ensure config is recreated atomically
We have parallel jobs such as kvm-update which consume
cves-autotriage.conf, if they run when this is being updates (as happens
at *:00 due to the relative cronjob timings) it could see a partial
configuration.  This leads it to ratchet down and retrigger old tags.
Install the new version of the configuration atomically.

Signed-off-by: Andy Whitcroft <apw@canonical.com>