mainline-builds: worker -- switch to using command groups
Currently we have one queue per command, this allows us to prioritise
instances of commands against each other but commands are still round-robin
against each other.  Switch commands into larger groups so that priorities
within the group are honoured.  If you listen to a group of command
you must be able to execute all of the possible commands in the group.
As part of this admin command become local only, but the local queue
continues to listen to all commands from all and any group we listen to.

Signed-off-by: Andy Whitcroft <apw@canonical.com>