mainline-builds: cod-incoming-publish -- revamp locking
Locking: we wish to allow multiple results to be pushed in parallel.
But once the publisher starts we want it to preempt any further uploaders
starting to push.  We use two lockfiles.  Exclusive lockers (publisher)
take an exclusive lock on A and then an exclusive lock on B.  Shared
lockers (uploaders) take an exclusive lock on A, then a shared lock on
lock on B, and then drop their lock on A.

Signed-off-by: Andy Whitcroft <apw@canonical.com>