mainline-builds: cod-incoming-publish -- record announcement flag on publication
We need to publish a result once it is complete.  To know we need to do
this we need to record that this publication is incomplete.  When
emitting a publication which starts children record the need to
announce; include information on whether it is the first build or a
rebuild.

Signed-off-by: Andy Whitcroft <apw@canonical.com>