mainline-builds: add latest-tarballs publisher
We are asked for simple tarballs of the trees representing the kernels
which we care to have CI done against.  Initially the ask is for
wireguard testing but the concept is generic.  Add support for
accumulating and maintaining a tarballs for the -proposed and -updates
kernels for all public releases which are not EOL.  Add a wireguard
specific index to drive which kernels to CI against.

Signed-off-by: Andy Whitcroft <apw@canonical.com>