cranky: update-dependent -- Use stricter matching for abi
The linux-uc20-efi packages have an extra .N as compared with
other kernel packages, e.g. 5.4.0-32.36.1. The abi is currently
found by chopping off the last .N from the version, which does
not work for these packages. Use a regex to provide stricter
matching against the part of the version string which specifies
the abi.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>