chroot-setup: switch chroot selection to the local architecture
We currently assume all builders are amd64 and have a fixed list of
chroots to build on a builder.  We should actually consider all build
architectures for the series, and whether this host can build for it.
Currently that means that if the host is amd64 then it can make amd64
and i386 chroots for the series; otherwise then it only would make a host
architecture chroot if that architecture is a listed.

Signed-off-by: Andy Whitcroft <apw@canonical.com>