adt-matrix: Include test packages from seed file in package list
We have a problem when testing a new package in a ppa. Britney
doesn't know about the package, so there is no data about the
package to populate the seeds in package-relation.json. Without
these seeds adt-matrix will not try to download any test results
for the package.

Since we already add baseline results for our packages in the
form of seed files in adt-matrix-hints, use this package list to
augment the package list from britney.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>