mainline-builds: cod-*-core -- propogate priority to tip builds
When we are submitted at a specific priority we want to submit our
tip builds at a corresponding priority.  We submit our tip builds
at the priority of the core builds minus one so that our tip builds
are prioritised behind other -core builds which we know are short.
This effectivly means we should consider priorities as bands two priority
levels apart.

Signed-off-by: Andy Whitcroft <apw@canonical.com>