announcer: add a new amqp announcer
We currently have to configure all of the possible machine which may
announce things to talk to irc, email, and now mattermost.  Add a new
amqp based announcer service which all of the workers etc can send their
announcements to.  That one machine will then use the various tools to
announce the item.

Signed-off-by: Andy Whitcroft <apw@canonical.com>