cranky: fix -- handle empty variant lists
The kernel-series.yaml provides variants information for each source.
This is a list of one or more variants we which expressed by that
source.  When absent we imply the '--' (primary empty) variant.

For some packages we wish to have variants and add-hoc additional names;
as an example we add -gke-5.3 meta-packages for just limited
architectures and flavours.  Over the lifetime of this source we shed
the variants to later packages (as the -gcp kernel rolls).  Eventually
we end up with no variants.  This is currently unspecified but is
logically represented by a variants specifier with a null list:

  variants: []

Currently a cranky fix takes the variants list and expresses it as
debian/variants within each package as it is cranked.  This considers an
absent variants list as the empty string.  This is mapped to a missing
debian/variants file and that in turn defaults the variants list to
'--'.  A null list is validly represented by a present debian/variants
file with no contents.  Fix shell-helper to differentiate the missing
and empty variant specifiers (missing is represented as ['--']).  Fix
the cranky variants fix to remove the file when variants are exactly
'--' ensuring we carry empty variants to the file.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>